pyproject.toml
[tool.poetry]
name = "dj_microservices"
version = "0.1.0"
description = ""
authors = ["surajmt8848 <suraj8848mt@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
djangorestframework = "^3.12.4"
django-cors-headers = "^3.7.0"
pika = "^1.2.0"
psycopg2-binary = "^2.8.6"
gunicorn = "^20.1.0"
Pillow = "^8.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
